pyproject.toml
[tool.poetry]
name = "drf-24-1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "drf_24"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
psycopg2 = "^2.9.9"
pillow = "^10.3.0"
python-dotenv = "^1.0.1"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
pytest-html = "^4.1.1"
coverage = "^7.5.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
